Label rectangles with letters instead of numbers
This is just a suggestion to make the presentation of the problem a little clearer.

I was momentarily confused with the two sets of numbers in the image of the puzzle description. It would be easier to understand if the rectangles were labeled with letters `A-K` instead of numbers `0-10`. However, that requires updating the image /:

If you don't feel it's worth the effort, I getcha.

Cheers! (cool puzzle, by the way)